pyproject.toml
[project]
name = "hf-data-check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets", 
    "torch", 
    "matplotlib", 
    "pillow", 
    "huggingface_hub"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["hello", "visualize_cc12m"]
